// src/api/generated/zod.ts
export const addEventApiV1LiabilitiesLiabilityIdEventsPostBodyAmountRegExpTwo = new RegExp('^(?!^[-+.]\*$)[+-]?0\*\\d\*\\.?\\d\*$');
export const addEventApiV1LiabilitiesLiabilityIdEventsPostBodyCurrencyMax = 16;

export const addEventApiV1LiabilitiesLiabilityIdEventsPostBodyPrincipalComponentMinOne = 0;

export const addEventApiV1LiabilitiesLiabilityIdEventsPostBodyPrincipalComponentRegExpTwo = new RegExp('^(?!^[-+.]\*$)[+-]?0\*\\d\*\\.?\\d\*$');
export const addEventApiV1LiabilitiesLiabilityIdEventsPostBodyInterestComponentMinOne = 0;

export const addEventApiV1LiabilitiesLiabilityIdEventsPostBodyInterestComponentRegExpTwo = new RegExp('^(?!^[-+.]\*$)[+-]?0\*\\d\*\\.?\\d\*$');


export const addEventApiV1LiabilitiesLiabilityIdEventsPostBody = zod.object({
  "type": zod.enum(['disbursement', 'repayment', 'interest']),
  "amount": zod.union([zod.number(),zod.string().regex(addEventApiV1LiabilitiesLiabilityIdEventsPostBodyAmountRegExpTwo)]),
  "currency": zod.string().min(1).max(addEventApiV1LiabilitiesLiabilityIdEventsPostBodyCurrencyMax),
  "occurred_at": zod.string().datetime({}),
  "principal_component": zod.union([zod.number().min(addEventApiV1LiabilitiesLiabilityIdEventsPostBodyPrincipalComponentMinOne),zod.string().regex(addEventApiV1LiabilitiesLiabilityIdEventsPostBodyPrincipalComponentRegExpTwo),zod.null()]).optional(),
  "interest_component": zod.union([zod.number().min(addEventApiV1LiabilitiesLiabilityIdEventsPostBodyInterestComponentMinOne),zod.string().regex(addEventApiV1LiabilitiesLiabilityIdEventsPostBodyInterestComponentRegExpTwo),zod.null()]).optional()
})